feat: remove console log from scroll collapse function; update header props import and refactor header stories for consistency
